// apps/api/src/max/schema.gen.ts
export interface components {
    schemas: {
// ...
        CallbackButton: Omit<components["schemas"]["Button"], "type"> & {
            /** @description Button payload */
            payload: string;
        } & {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "callback";
        };
        /** @description After pressing this type of button user follows the link it contains */
        LinkButton: Omit<components["schemas"]["Button"], "type"> & {
            url: string;
        } & {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "link";
        };
// ...
        RequestContactButton: Omit<components["schemas"]["Button"], "type"> & {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "request_contact";
        };
// ...
        OpenAppButton: Omit<components["schemas"]["Button"], "type"> & {
            /** @description Button payload */
            payload?: string | null;
            /** @description Unique public name of the bot wired to the mini app */
            web_app: string;
            /** @description Unique identifier of the bot wired to the mini app */
            contact_id?: components["schemas"]["UserId"] | null;
        } & {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "open_app";
        };
        /** @description After pressing this type of button client copies payload data to clipboard */
        ClipboardButton: Omit<components["schemas"]["Button"], "type"> & {
            /** @description Button payload */
            payload: string;
        } & {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "clipboard";
        };
// ...
    };
// ...
}